Implement getAllStory with userId - Implement ControllerAdvice - Rename package `upload` to `storage`
