+++ src/main/kotlin/com/trip/my/room/server/common/storage/StorageProperties.kt
@@ -1,4 +1,4 @@
-package com.trip.my.room.server.common.upload
+package com.trip.my.room.server.common.storage
 
 import java.util.*
